Add Utils to convert PS XML to JSON
* Use `xml2js` for converting XML to JSON
* Add Utility functions to further format the JSON to requrements
* Add tests for all utility functions

After considering the many libraries to convert XML to JSON, settled
on `xml2js` for the following reasons:

1. Need to ignore namespaces and attributes
2. `xml2js` is popular with hundreds of thousands of downloads a month

One caveat is that none of the libraries allowed for custom functions to
modify the JSON structure - only values and names.

This is required as PS XML contains response of type:

```xml
<ExampleArray>
  <Example>hello</Example>
</ExampleArray>
```

Which gets converted to:
```javascript
{
  "ExampleArray": {
    "Example": "hello"
  }
}
```

Whereas, what's expected is:
```
{
  "ExampleArray": ["hello"]
}
```

Also, all the blindly created arrays if multiple nodes existed with the
same name.